refactor attempt
