+++ src/components/faqTemplates/shared/index.jsx
@@ -1,0 +1,3 @@
+import { ExpandableFaqList } from './FaqList'
+
+export { ExpandableFaqList }
